Compiles without errors or warnings pt 2.
